Develop (#15)
* units tests (2)

* units tests (2)

* more unit tests (haikuService)

* more unit tests (haikuService)

* more unit tests (haikuService)

* more unit tests (haikuService)

+++ src/main/java/com/jux/familyspace/api/ElementSizeTrackerInterface.java
@@ -3,9 +3,9 @@ package com.jux.familyspace.api;
 public interface ElementSizeTrackerInterface<T> {
 
 
-    int getTotalSize();
+    long getTotalSize();
 
-    void setTotalSize(int size);
+    void setTotalSize(long size);
 
     void incrementSize(int delta);
