+++ src/main/java/com/jux/familyspace/component/DailyThoughtSizeTracker.java
@@ -4,20 +4,20 @@ import com.jux.familyspace.api.ElementSizeTrackerInterface;
 import com.jux.familyspace.model.DailyThought;
 import org.springframework.stereotype.Component;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Component
 public class DailyThoughtSizeTracker implements ElementSizeTrackerInterface<DailyThought> {
 
-    private final AtomicInteger totalDailyThoughtsSize = new AtomicInteger();
+    private final AtomicLong totalDailyThoughtsSize = new AtomicLong();
 
     @Override
-    public int getTotalSize() {
+    public long getTotalSize() {
         return totalDailyThoughtsSize.get();
     }
 
     @Override
-    public void setTotalSize(int size) {
+    public void setTotalSize(long size) {
         totalDailyThoughtsSize.set(size);
     }
